Moved matter_test into matter.rs
